Create xts addfactor

Create xts addfactor

Usage

addf(
  start = bnk_start,
  end = bnk_end,
  from = 0,
  to = 0,
  ser_name = "value",
  per = "year"
)

Arguments

start

start date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")

end

end date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")

from

first value for linear interpolation (numeric)

to

last value for linear interpolation (numeric)

ser_name

name of the xts series (string)

per

periodicity of series (character: "year" - default) if date format of start is quarterly, automatically set to "quarter"

Value

a single xts series spanning bnk_start-bnk_end

Details

this is a wrapper around make_xts with some additional functionality. the start and end dates specify the span of the non-zero add-factor value. the remaining period between start and end is filled with zeros.

Examples

addf()
#>            value
#> 1970-01-01     0
#> 1971-01-01     0
#> 1972-01-01     0
#> 1973-01-01     0
#> 1974-01-01     0
#> 1975-01-01     0
#> 1976-01-01     0
#> 1977-01-01     0
#> 1978-01-01     0
#> 1979-01-01     0
#> 1980-01-01     0
#> 1981-01-01     0
#> 1982-01-01     0
#> 1983-01-01     0
#> 1984-01-01     0
#> 1985-01-01     0
#> 1986-01-01     0
#> 1987-01-01     0
#> 1988-01-01     0
#> 1989-01-01     0
#> 1990-01-01     0
#> 1991-01-01     0
#> 1992-01-01     0
#> 1993-01-01     0
#> 1994-01-01     0
#> 1995-01-01     0
#> 1996-01-01     0
#> 1997-01-01     0
#> 1998-01-01     0
#> 1999-01-01     0
#> 2000-01-01     0
#> 2001-01-01     0
#> 2002-01-01     0
#> 2003-01-01     0
#> 2004-01-01     0
#> 2005-01-01     0
#> 2006-01-01     0
#> 2007-01-01     0
#> 2008-01-01     0
#> 2009-01-01     0
#> 2010-01-01     0
#> 2011-01-01     0
#> 2012-01-01     0
#> 2013-01-01     0
#> 2014-01-01     0
#> 2015-01-01     0
#> 2016-01-01     0
#> 2017-01-01     0
#> 2018-01-01     0
#> 2019-01-01     0
#> 2020-01-01     0
#> 2021-01-01     0
#> 2022-01-01     0
#> 2023-01-01     0
#> 2024-01-01     0
#> 2025-01-01     0
#> 2026-01-01     0
#> 2027-01-01     0
#> 2028-01-01     0
#> 2029-01-01     0
#> 2030-01-01     0
#> 2031-01-01     0
#> 2032-01-01     0
#> 2033-01-01     0
#> 2034-01-01     0
#> 2035-01-01     0
#> 2036-01-01     0
#> 2037-01-01     0
#> 2038-01-01     0
#> 2039-01-01     0
#> 2040-01-01     0
#> 2041-01-01     0
#> 2042-01-01     0
#> 2043-01-01     0
#> 2044-01-01     0
#> 2045-01-01     0
#> 2046-01-01     0
#> 2047-01-01     0
#> 2048-01-01     0
#> 2049-01-01     0
#> 2050-01-01     0
#> 2051-01-01     0
#> 2052-01-01     0
#> 2053-01-01     0
#> 2054-01-01     0
#> 2055-01-01     0
#> 2056-01-01     0
#> 2057-01-01     0
#> 2058-01-01     0
#> 2059-01-01     0
#> 2060-01-01     0
addf(201002, 201504, 1, 2)
#>            value
#> 1970-01-01     0
#> 1970-04-01     0
#> 1970-07-01     0
#> 1970-10-01     0
#> 1971-01-01     0
#> 1971-04-01     0
#> 1971-07-01     0
#> 1971-10-01     0
#> 1972-01-01     0
#> 1972-04-01     0
#>        ...      
#> 2058-07-01     0
#> 2058-10-01     0
#> 2059-01-01     0
#> 2059-04-01     0
#> 2059-07-01     0
#> 2059-10-01     0
#> 2060-01-01     0
#> 2060-04-01     0
#> 2060-07-01     0
#> 2060-10-01     0
addf(20100101, 20601201, 1, 2, per = "month")
#>               value
#> 1970-01-01 0.000000
#> 1970-02-01 0.000000
#> 1970-03-01 0.000000
#> 1970-04-01 0.000000
#> 1970-05-01 0.000000
#> 1970-06-01 0.000000
#> 1970-07-01 0.000000
#> 1970-08-01 0.000000
#> 1970-09-01 0.000000
#> 1970-10-01 0.000000
#>        ...         
#> 2060-03-01 1.985270
#> 2060-04-01 1.986907
#> 2060-05-01 1.988543
#> 2060-06-01 1.990180
#> 2060-07-01 1.991817
#> 2060-08-01 1.993453
#> 2060-09-01 1.995090
#> 2060-10-01 1.996727
#> 2060-11-01 1.998363
#> 2060-12-01 2.000000
addf(20100101, from = 1, to = 2, per = "quarter")
#>               value
#> 1970-01-01 0.000000
#> 1970-04-01 0.000000
#> 1970-07-01 0.000000
#> 1970-10-01 0.000000
#> 1971-01-01 0.000000
#> 1971-04-01 0.000000
#> 1971-07-01 0.000000
#> 1971-10-01 0.000000
#> 1972-01-01 0.000000
#> 1972-04-01 0.000000
#>        ...         
#> 2058-07-01 1.955665
#> 2058-10-01 1.960591
#> 2059-01-01 1.965517
#> 2059-04-01 1.970443
#> 2059-07-01 1.975369
#> 2059-10-01 1.980296
#> 2060-01-01 1.985222
#> 2060-04-01 1.990148
#> 2060-07-01 1.995074
#> 2060-10-01 2.000000
addf(2010.2, 2015.4, 1, 2, "ECT_HI")
#>            ECT_HI
#> 1970-01-01      0
#> 1970-04-01      0
#> 1970-07-01      0
#> 1970-10-01      0
#> 1971-01-01      0
#> 1971-04-01      0
#> 1971-07-01      0
#> 1971-10-01      0
#> 1972-01-01      0
#> 1972-04-01      0
#>        ...       
#> 2058-07-01      0
#> 2058-10-01      0
#> 2059-01-01      0
#> 2059-04-01      0
#> 2059-07-01      0
#> 2059-10-01      0
#> 2060-01-01      0
#> 2060-04-01      0
#> 2060-07-01      0
#> 2060-10-01      0