Fiscal year to date growth rate

Fiscal year to date growth rate

Usage

fytd_gr(x)

Arguments

x

a ts-boxable object

Value

object of the same type as the input containing year to date growth rate

Details

this function operates similarly to ytd_gr() but assumes that the fiscal year starts in July and accumulates the values from that month onward.

Examples

quarterly_data_example |>
  fytd_gr()
#> # A tibble: 102 × 31
#>    time       E_NF_HI ECT_HI  EMN_HI EWT_HI ERT_HI E_TRADE_HI E_TU_HI ETWTANS_HI
#>    <date>       <dbl>  <dbl>   <dbl>  <dbl>  <dbl>      <dbl>   <dbl>      <dbl>
#>  1 1998-01-01  NA     NA     NA      NA     NA         NA      NA         NA    
#>  2 1998-04-01  NA     NA     NA      NA     NA         NA      NA         NA    
#>  3 1998-07-01  NA     NA     NA      NA     NA         NA      NA         NA    
#>  4 1998-10-01  NA     NA     NA      NA     NA         NA      NA         NA    
#>  5 1999-01-01   0.136 -1.35  -0.692   0.126 -1.54      -1.19   -2.06      -0.667
#>  6 1999-04-01   0.163 -0.865 -0.0474 -0.157 -0.886     -0.730  -1.90      -0.500
#>  7 1999-07-01   0.871 -0.304  2.09   -0.440  1.40       1.02   -0.145      0.990
#>  8 1999-10-01   1.36   0.413  2.60   -0.125  2.05       1.60    1.00       1.99 
#>  9 2000-01-01   1.84   2.41   2.79    0.881  2.38       2.08    1.81       3.69 
#> 10 2000-04-01   2.22   4.14   2.88    1.29   2.26       2.07    1.99       4.27 
#> # ℹ 92 more rows
#> # ℹ 22 more variables: ETWNS_HI <dbl>, EUT_HI <dbl>, EIF_HI <dbl>,
#> #   EFI_HI <dbl>, E_FIR_HI <dbl>, ERE_HI <dbl>, EPS_HI <dbl>, E_PBS_HI <dbl>,
#> #   E_ELSE_HI <dbl>, EMA_HI <dbl>, EAD_HI <dbl>, EED_HI <dbl>, EHC_HI <dbl>,
#> #   EAF_HI <dbl>, EAFAC_HI <dbl>, EAFFD_HI <dbl>, EOS_HI <dbl>, EGV_HI <dbl>,
#> #   EGVFD_HI <dbl>, EGVST_HI <dbl>, EGVSTEDNS_HI <dbl>, EGVLC_HI <dbl>
monthly_data_example |>
  tsbox::ts_long() |>
  tsbox::ts_pick("VISNS_HI") |>
  tsbox::ts_xts() |>
  fytd_gr() |>
  tail()
#>             VISNS_HI
#> 2023-02-01 21.809194
#> 2023-03-01 20.743637
#> 2023-04-01 18.394541
#> 2023-05-01 16.855794
#> 2023-06-01 15.732838
#> 2023-07-01  1.534944