Interactive plot with level and growth rate for forecast series

Interactive plot with level and growth rate for forecast series

Usage

plot_fc(
  ser,
  rng_start = as.character(Sys.Date() - lubridate::years(15)),
  rng_end = as.character(Sys.Date()),
  height = 300,
  width = 900,
  yoy_gr = TRUE
)

Arguments

ser

time series to plot (min 1, max 3) (e.g. current fcst, old fcst, history)

rng_start

start of zoom range ("YYYY-MM-DD")

rng_end

end of the zoom range ("YYYY-MM-DD")

height

height of a single panel (px)

width

width of a single panel (px)

yoy_gr

year-over-year (default) or annualized growth

Value

a dygraph plot

Examples

`ncen@us.sola` <- ts(NA_real_, start = 2016, end = 2021, freq = 1) |> tsbox::ts_xts()
`ncen@us.sola`["2016/2021"] <- c(323127513, 325511184, 327891911, 330268840, 332639102, 334998398)
`ncen@us.oldsola` <- `ncen@us.sola`
`ncen@us.oldsola`["2020/2021"] <- c(352639102, 374998398)
`ncen@us.a` <- `ncen@us.sola`
`ncen@us.a`["2020/2021"] <- NA
test1 <- tsbox::ts_tslist(tsbox::ts_c(`ncen@us.sola`, `ncen@us.oldsola`, `ncen@us.a`)) |>
  purrr::map(AtoQ) |>
  purrr::reduce(tsbox::ts_c) |>
  magrittr::set_names(c("ncen@us.sola", "ncen@us.oldsola", "ncen@us.a"))
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
#> Joining with `by = join_by(time)`
plot_fc(tsbox::ts_c(`ncen@us.sola`, `ncen@us.oldsola`, `ncen@us.a`), rng_start = "2017-01-01")

plot_fc(test1, rng_start = "2017-01-01")

test1 <- AtoQ(`ncen@us.sola`)
#> Joining with `by = join_by(time)`
plot_fc(`ncen@us.sola`, rng_start = "2017-01-01")

plot_fc(test1, rng_start = "2017-01-01")

if (FALSE) { # interactive()
get_series_exp(74, rename = "no") |>
  tsbox::ts_long() |>
  tsbox::ts_xts() |>
  magrittr::extract(, c("E_NF@HI.Q", "ECT@HI.Q", "EMN@HI.Q")) |>
  plot_fc()
}