Create xts addfactor

Create xts addfactor

Usage

addf(
  start = bnk_start,
  end = bnk_end,
  from = 0,
  to = 0,
  ser_name = "value",
  per = "year"
)

Arguments

start

start date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")

end

end date of linear interpolation (character: "yyyy-mm-dd", "yyyyqq", "yyyy")

from

first value for linear interpolation (numeric)

to

last value for linear interpolation (numeric)

ser_name

name of the xts series (string)

per

periodicity of series (character: "year" - default) if date format of start is quarterly, automatically set to "quarter"

Value

a single xts series spanning bnk_start-bnk_end

Details

this is a wrapper around make_xts with some additional functionality. the start and end dates specify the span of the non-zero add-factor value. the remaining period between start and end is filled with zeros.

Examples

addf()
#>            value
#> 1900-01-01     0
#> 1901-01-01     0
#> 1902-01-01     0
#> 1903-01-01     0
#> 1904-01-01     0
#> 1905-01-01     0
#> 1906-01-01     0
#> 1907-01-01     0
#> 1908-01-01     0
#> 1909-01-01     0
#>        ...      
#> 2051-01-01     0
#> 2052-01-01     0
#> 2053-01-01     0
#> 2054-01-01     0
#> 2055-01-01     0
#> 2056-01-01     0
#> 2057-01-01     0
#> 2058-01-01     0
#> 2059-01-01     0
#> 2060-01-01     0
addf(201002, 201504, 1, 2)
#>            value
#> 1900-01-01     0
#> 1901-01-01     0
#> 1902-01-01     0
#> 1903-01-01     0
#> 1904-01-01     0
#> 1905-01-01     0
#> 1906-01-01     0
#> 1907-01-01     0
#> 1908-01-01     0
#> 1909-01-01     0
#>        ...      
#> 2051-01-01     0
#> 2052-01-01     0
#> 2053-01-01     0
#> 2054-01-01     0
#> 2055-01-01     0
#> 2056-01-01     0
#> 2057-01-01     0
#> 2058-01-01     0
#> 2059-01-01     0
#> 2060-01-01     0
addf(20100101, 20601201, 1, 2, per = "month")
#>               value
#> 1900-01-01 0.000000
#> 1900-02-01 0.000000
#> 1900-03-01 0.000000
#> 1900-04-01 0.000000
#> 1900-05-01 0.000000
#> 1900-06-01 0.000000
#> 1900-07-01 0.000000
#> 1900-08-01 0.000000
#> 1900-09-01 0.000000
#> 1900-10-01 0.000000
#>        ...         
#> 2060-03-01 1.985270
#> 2060-04-01 1.986907
#> 2060-05-01 1.988543
#> 2060-06-01 1.990180
#> 2060-07-01 1.991817
#> 2060-08-01 1.993453
#> 2060-09-01 1.995090
#> 2060-10-01 1.996727
#> 2060-11-01 1.998363
#> 2060-12-01 2.000000
addf(20100101, from = 1, to = 2, per = "quarter")
#>               value
#> 1900-01-01 0.000000
#> 1901-01-01 0.000000
#> 1902-01-01 0.000000
#> 1903-01-01 0.000000
#> 1904-01-01 0.000000
#> 1905-01-01 0.000000
#> 1906-01-01 0.000000
#> 1907-01-01 0.000000
#> 1908-01-01 0.000000
#> 1909-01-01 0.000000
#>        ...         
#> 2058-07-01 1.955665
#> 2058-10-01 1.960591
#> 2059-01-01 1.965517
#> 2059-04-01 1.970443
#> 2059-07-01 1.975369
#> 2059-10-01 1.980296
#> 2060-01-01 1.985222
#> 2060-04-01 1.990148
#> 2060-07-01 1.995074
#> 2060-10-01 2.000000
addf(2010.2, 2015.4, 1, 2, "ECT_HI")
#>            ECT_HI
#> 1900-01-01      0
#> 1901-01-01      0
#> 1902-01-01      0
#> 1903-01-01      0
#> 1904-01-01      0
#> 1905-01-01      0
#> 1906-01-01      0
#> 1907-01-01      0
#> 1908-01-01      0
#> 1909-01-01      0
#>        ...       
#> 2051-01-01      0
#> 2052-01-01      0
#> 2053-01-01      0
#> 2054-01-01      0
#> 2055-01-01      0
#> 2056-01-01      0
#> 2057-01-01      0
#> 2058-01-01      0
#> 2059-01-01      0
#> 2060-01-01      0